update[rombel][walikelas]:menambahkan walikelas pada rombel dan membuat dashboard dan data siswa pada role walikelas

+++ Modules/Pkl/app/Services/Data/SiswaService.php
@@ -3,9 +3,11 @@
 namespace Modules\Pkl\Services\Data;
 
 use App\Models\Jurusan;
+use App\Models\Rombel;
 use App\Services\DataTableService;
 use Exception;
 use Illuminate\Database\QueryException;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Cache;
 use Illuminate\Support\Facades\Log;
 use Modules\Pkl\Repositories\SiswaRepository;
@@ -141,10 +143,17 @@ class SiswaService
      *
      * @return mixed Data dalam format JSON untuk DataTable.
      */
-    public function table()
+    public function table($role = null)
     {
-        return DataTableService::draw('siswas')
-            ->where('deleted_at', null)
+        $query = DataTableService::draw('siswas');
+
+        if ($role == 'wali_kelas') {
+            $user = Auth::user();
+            $getRombel = Rombel::where('walikelas_id', $user->id)->first();
+            $query->where('rombel_id', $getRombel->id);
+        }
+
+        return $query->where('deleted_at', null)
             ->addColumn('jurusan', function ($detail) {
                 return Cache::remember("jurusan_{$detail->jurusan_id}", now()->addMinutes(5), function () use ($detail) {
                     return Jurusan::find($detail->jurusan_id)->name ?? 'Unknown';
@@ -164,21 +173,38 @@ class SiswaService
                         </div>
                 ';
             })
-            ->addColumn('action', function ($detail) {
-                return '
-                <div class="d-inline-block">
-                    <a href="javascript:void(0);" class="btn btn-sm rounded-pill btn-icon dropdown-toggle hide-arrow show" data-bs-toggle="dropdown" aria-expanded="true"><i class="ti ti-dots-vertical ti-md"></i></a>
-                    <ul class="dropdown-menu dropdown-menu-end m-0" data-popper-placement="bottom-end">
-                        <li>
-                            <a class="dropdown-item" href="javascript:void(0);" data-permision="user-update" onclick="editData(this)" data-params="' . base64_encode(json_encode($detail)) . '">Edit</a>
-                        </li>
-                        <div class="dropdown-divider"></div>
-                        <li>
-                            <a class="dropdown-item" href="javascript:void(0);" data-permision="user-update" onclick="deleteData(this)" data-params="' . base64_encode(json_encode($detail)) . '">Delete</a>
-                        </li>
-                    </ul>
-                </div>
-                ';
+            ->addColumn('action', function ($detail) use ($role) {
+                if($role == 'wali_kelas'){
+                    return '
+                    <div class="d-flex align-items-center">
+                        <a class="btn btn-icon" href="javascript:void(0);" data-permision="user-update" onclick="onEdit(this)" data-params="' . base64_encode(json_encode($detail)) . '">
+                            <i class="ti ti-eye ti-md"></i>
+                        </a>
+                        <div class="dropdown">
+                            <a href="javascript:;" class="btn dropdown-toggle hide-arrow btn-icon p-0" data-bs-toggle="dropdown" aria-expanded="false"><i class="ti ti-dots-vertical ti-md"></i></a>
+                            <div class="dropdown-menu dropdown-menu-end">
+                                <a class="dropdown-item" href="javascript:void(0);" data-permision="user-update" onclick="onDetails(this)" data-params="' . base64_encode(json_encode($detail)) . '">Reset Password</a>
+                            </div>
+                        </div>
+                    </div>
+                    ';
+                }else{
+                    return '
+                    <div class="d-inline-block">
+                        <a href="javascript:void(0);" class="btn btn-sm rounded-pill btn-icon dropdown-toggle hide-arrow show" data-bs-toggle="dropdown" aria-expanded="true"><i class="ti ti-dots-vertical ti-md"></i></a>
+                        <ul class="dropdown-menu dropdown-menu-end m-0" data-popper-placement="bottom-end">
+                            <li>
+                                <a class="dropdown-item" href="javascript:void(0);" data-permision="user-update" onclick="editData(this)" data-params="' . base64_encode(json_encode($detail)) . '">Edit</a>
+                            </li>
+                            <div class="dropdown-divider"></div>
+                            <li>
+                                <a class="dropdown-item" href="javascript:void(0);" data-permision="user-update" onclick="deleteData(this)" data-params="' . base64_encode(json_encode($detail)) . '">Delete</a>
+                            </li>
+                        </ul>
+                    </div>
+                    ';
+                }
+
             })
             ->rawColumns(['status', 'action'])
             ->toJson();
